Fit univariate latent change score models

Fit univariate latent change score models.

fit_uni_lcsm(data, var, model, mimic = "Mplus", estimator = "MLR",
  missing = "FIML", return_lavaan_syntax = FALSE,
  return_lavaan_syntax_string = FALSE, ...)

Arguments

data
Wide dataset.

var
List of variables measuring construct of the model.

model
List of model specifications (logical) for variables specified in var.

alpha_constant (Constant change factor),

alpha_piecewise (Piecewise constant change factors),

alpha_piecewise_num (Changepoint of piecewise constant change factors,

alpha_linear (Linear change factor),

beta (Proportional change factor),

phi (Autoregression of change scores).

mimic
See lavaan.

estimator
See lavaan.

missing
See lavaan.

return_lavaan_syntax
Logical, if TRUE return the lavaan syntax used for simulating data, looking beautiful using cat

return_lavaan_syntax_string
Logical, if return_lavaan_syntax is TRUE and return_lavaan_syntax_string == TRUE return the lavaan syntax as one ugly string

...
Additional arguments to be passed to lavaan.

Value

This function returns a lavaan class object.

References

Ghisletta, P., & McArdle, J. J. (2012). Latent Curve Models and Latent Change Score Models Estimated in R. Structural Equation Modeling: A Multidisciplinary Journal, 19(4), 651–682. https://doi.org/10.1080/10705511.2012.713275.

Grimm, K. J., Ram, N., & Estabrook, R. (2017). Growth Modeling—Structural Equation and Multilevel Modeling Approaches. New York: The Guilford Press.

McArdle, J. J. (2009). Latent variable modeling of differences and changes with longitudinal data. Annual Review of Psychology, 60(1), 577–605. https://doi.org/10.1146/annurev.psych.60.110707.16361.

Yves Rosseel (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1-36. http://www.jstatsoft.org/v48/i02/.

Examples

# Fit univariate latent change score model
fit_uni_lcsm(data = data_uni_lcsm, var = names(data_uni_lcsm)[-1],
             model = list(alpha_constant = TRUE, beta = FALSE, phi = FALSE))

#> lavaan 0.6-5 ended normally after 56 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                         15
#>   Number of equality constraints                     9
#>   Row rank of the constraints matrix                 9
#>                                                       
#>   Number of observations                           500
#>   Number of missing patterns                       273
#>                                                       
#> Model Test User Model:
#>                                               Standard      Robust
#>   Test Statistic                                76.647      75.722
#>   Degrees of freedom                                59          59
#>   P-value (Chi-square)                           0.061       0.070
#>   Scaling correction factor                                  1.012
#>     for the Yuan-Bentler correction (Mplus variant)