Simulate data from a univariate latent change score model

TODO: Describe function

sim_uni_lcsm(
  timepoints,
  model,
  model_param = NULL,
  var = "x",
  change_letter = "g",
  sample.nobs = 500,
  na_pct = 0,
  ...,
  return_lavaan_syntax = FALSE,
  return_lavaan_syntax_string = FALSE
)

Arguments

timepoints
See specify_uni_lcsm

model
See specify_uni_lcsm

model_param
List, specifying parameter estimates for the LCS model that has been specified in the argument 'model'

gamma_lx1: Mean of latent true scores x (Intercept),

sigma2_lx1: Variance of latent true scores x,

sigma2_ux: Variance of observed scores x,

alpha_g2: Mean of change factor (g2),

alpha_g3: Mean of change factor (g3),

sigma2_g2: Variance of constant change factor (g2).

sigma2_g3: Variance of constant change factor (g3),

sigma_g2lx1: Covariance of constant change factor (g2) with the initial true score x (lx1),

sigma_g3lx1: Covariance of constant change factor (g3) with the initial true score x (lx1),

sigma_g2g3: Covariance of change factors (g2 and g2),

phi_x: Autoregression of change scores x.

var
See specify_uni_lcsm

change_letter
See specify_uni_lcsm

sample.nobs
Numeric, number of cases to be simulated, see specify_uni_lcsm

na_pct
Numeric, percentage of random missing values in the simulated dataset [0,1]

...
Arguments to be passed on to simulateData

return_lavaan_syntax
Logical, if TRUE return the lavaan syntax used for simulating data, looking beautiful using cat. Note that this output cant be saved as an object. To save the lavaan syntax as an object the argument 'return_lavaan_syntax_string' has to be set to TRUE.

return_lavaan_syntax_string
Logical, if 'return_lavaan_syntax' = TRUE and 'return_lavaan_syntax_string' = TRUE return the lavaan syntax as one ugly string

Value

tibble

Examples

# Simulate data from univariate LCS model parameters 
sim_uni_lcsm(timepoints = 10,
             model = list(alpha_constant = TRUE, beta = FALSE, phi = TRUE),
             model_param = list(gamma_lx1 = 21,
                                sigma2_lx1 = 1.5,
                                sigma2_ux = .2,
                                alpha_j2 = -.93,
                                sigma2_j2 = .1,
                                sigma_j2lx1 = .2),
             return_lavaan_syntax = FALSE,
             return_lavaan_syntax_string = FALSE,
             sample.nobs = 1000,
             na_pct = .3)

#> Parameter estimates for the data simulation are taken from the argument 'model_param'.

#> Warning: The following parameters are specified in LCS model but no parameter estimates have been entered in 'model_param':
#> -  alpha_g2
#> -  sigma2_g2
#> -  sigma_g2lx1
#> -  phi_x

#> Warning: lavaan WARNING: some regression coefficients are unspecified and will be set to zero

#> # A tibble: 1,000 x 11
#>       id    x1    x2    x3    x4    x5    x6    x7    x8    x9   x10
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1     1  21.6  21.5  20.7  NA    NA    18.2  NA    17.6  17.1  16.5
#>  2     2  20.3  20.4  21.2  NA    22.4  23.5  23.5  23.9  23.6  24.2
#>  3     3  NA    19.5  18.2  19.0  NA    18.9  NA    17.8  18.0  17.5
#>  4     4  NA    NA    20.0  NA    20.3  19.5  19.8  20.4  NA    20.8
#>  5     5  20.2  18.8  NA    NA    NA    15.6  14.2  12.7  NA    11.9
#>  6     6  20.8  NA    20.5  NA    21.1  20.7  NA    19.3  19.4  19.7
#>  7     7  20.9  21.3  21.1  22.2  NA    NA    23.7  23.9  NA    25.2
#>  8     8  18.2  17.4  NA    16.8  15.6  15.4  14.4  13.0  NA    12.9
#>  9     9  19.9  19.8  NA    18.7  NA    18.1  18.1  17.7  NA    17.5
#> 10    10  20.9  20.5  23.6  24.4  NA    NA    29.6  29.6  32.0  33.6
#> # … with 990 more rows