Simulate data from univariate latent change score model parameter estimates

This function simulate data from univariate latent change score model parameter estimates using simulateData.

sim_uni_lcsm(
  timepoints,
  model,
  model_param = NULL,
  var = "x",
  change_letter = "g",
  sample.nobs = 500,
  na_pct = 0,
  ...,
  return_lavaan_syntax = FALSE
)

Arguments

timepoints
See specify_uni_lcsm

model
See specify_uni_lcsm

model_param
List, specifying parameter estimates for the LCS model that has been specified in the argument 'model'

gamma_lx1: Mean of latent true scores x (Intercept),

sigma2_lx1: Variance of latent true scores x,

sigma2_ux: Variance of observed scores x,

alpha_g2: Mean of change factor (g2),

alpha_g3: Mean of change factor (g3),

sigma2_g2: Variance of constant change factor (g2).

sigma2_g3: Variance of constant change factor (g3),

sigma_g2lx1: Covariance of constant change factor (g2) with the initial true score x (lx1),

sigma_g3lx1: Covariance of constant change factor (g3) with the initial true score x (lx1),

sigma_g2g3: Covariance of change factors (g2 and g2),

phi_x: Autoregression of change scores x.

var
See specify_uni_lcsm

change_letter
See specify_uni_lcsm

sample.nobs
Numeric, number of cases to be simulated, see specify_uni_lcsm

na_pct
Numeric, percentage of random missing values in the simulated dataset [0,1]

...
Arguments to be passed on to simulateData

return_lavaan_syntax
Logical, if TRUE return the lavaan syntax used for simulating data. To make it look beautiful use the function cat.

Value

tibble

Examples

# Simulate data from univariate LCS model parameters 
sim_uni_lcsm(timepoints = 10, 
             model = list(alpha_constant = TRUE, beta = FALSE, phi = TRUE), 
             model_param = list(gamma_lx1 = 21, 
                                sigma2_lx1 = 1.5,
                                sigma2_ux = .2, 
                                alpha_j2 = -.93,
                                sigma2_j2 = .1,
                                sigma_j2lx1 = .2),
             return_lavaan_syntax = FALSE, 
             sample.nobs = 1000,
             na_pct = .3)

#> Parameter estimates for the data simulation are taken from the argument 'model_param'.

#> Warning: The following parameters are specified in LCS model but no parameter estimates have been entered in 'model_param':
#> -  alpha_g2
#> -  sigma2_g2
#> -  sigma_g2lx1
#> -  phi_x

#> Warning: lavaan WARNING: some regression coefficients are unspecified and will be set to zero

#> # A tibble: 1,000 x 11
#>       id    x1    x2    x3    x4    x5    x6    x7    x8    x9   x10
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1     1  NA    21.1  NA    24.0  NA    28.3  NA    30.0  NA    NA  
#>  2     2  22.2  NA    NA    20.1  19.6  19.7  NA    19.3  18.4  18.4
#>  3     3  19.4  19.1  NA    NA    16.8  NA    16.1  15.3  14.5  13.1
#>  4     4  NA    NA    20.9  19.8  18.7  19.0  18.5  18.3  NA    17.4
#>  5     5  NA    20.1  NA    20.9  NA    20.9  NA    NA    21.2  22.4
#>  6     6  NA    NA    22.3  NA    22.8  NA    24.2  24.0  25.2  NA  
#>  7     7  21.3  21.0  19.9  20.2  18.7  NA    17.7  16.2  15.6  NA  
#>  8     8  20.2  20.4  18.1  17.6  16.6  15.8  14.8  NA    12.5  NA  
#>  9     9  NA    NA    24.9  25.5  NA    28.7  29.8  31.1  33.2  34.5
#> 10    10  NA    NA    NA    19.0  NA    18.2  17.4  16.6  15.1  14.0
#> # … with 990 more rows